Arc-drawing code, taken from cairosvg (Python)
The "Church" test files test drawing arcs.

I've also added note on copying the arc code from the Python in the
documentation.

+++ t/image-cairosvg.t
@@ -16,22 +16,30 @@ my $cairosvg2 = Image::CairoSVG->new (
 );
 ok ($cairosvg2);
 
-my $stem = "$FindBin::Bin/Technical_college";
-my $file = "$stem.svg";
-my $testpng = "$stem.png";
 
-$cairosvg2->render ($file);
-my $tempout = "$FindBin::Bin/TC-out.png";
-$surface->write_to_png ($tempout);
 SKIP: {
     eval {
 	require Image::PNG::Libpng;
     };
-    skip "No Image::PNG::Libpng", 1 if $@;
-    my $diff = Image::PNG::Libpng::image_data_diff ($testpng, $tempout);
-    ok (! $diff, "PNG files contain the same data");
-};
-unlink $tempout;
+    skip "No Image::PNG::Libpng", 2 if $@;
+
+    for my $f (qw/Technical_college Church/) {
+	my $surface = Cairo::ImageSurface->create ('argb32', 400, 400);
+	my $cairosvg2 = Image::CairoSVG->new (
+	    surface => $surface,
+	);
+	my $stem = "$FindBin::Bin/$f";
+	my $file = "$stem.svg";
+	my $testpng = "$stem.png";
+	die unless -f $testpng;
+	$cairosvg2->render ($file);
+	my $tempout = "$stem-out.png";
+	$surface->write_to_png ($tempout);
+	my $diff = Image::PNG::Libpng::image_data_diff ($testpng, $tempout);
+	ok (! $diff, "PNG files contain the same data");
+	unlink $tempout;
+    };
+}
 
 done_testing ();
